ovsdb: Fix -Wimplicit-int-float-conversion warning
This patch fixes the -Wimplicit-int-float-conversion report
by clang, and thus resolves the breakage of BSD build by
cirrus-ci. From:

https://cirrus-ci.com/task/5928699675541504?command=build#L1652

ovsdb/execution.c:733:32: error: implicit conversion from
'long long' to 'double' changes value from 9223372036854775807
to 9223372036854775808 [-Werror,-Wimplicit-int-float-conversion]
            timeout_msec = MIN(LLONG_MAX, json_real(timeout));
                           ~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/usr/include/sys/limits.h:69:19: note: expanded from macro 'LLONG_MAX'
                        ^~~~~~~~~~~
/usr/include/x86/_limits.h:74:21: note: expanded from macro '__LLONG_MAX'

+++ ovsdb/execution.c
@@ -730,7 +730,7 @@ ovsdb_execute_wait(struct ovsdb_execution *x, struct ovsdb_parser *parser,
     }
     if (!error) {
         if (timeout) {
-            timeout_msec = MIN(LLONG_MAX, json_real(timeout));
+            timeout_msec = MIN(LLONG_MAX, (long long int) json_real(timeout));
             if (timeout_msec < 0) {
                 error = ovsdb_syntax_error(timeout, NULL,
                                            "timeout must be nonnegative");
